+++ Common/src/main/java/com/almostreliable/unified/config/DuplicationConfig.java
@@ -1,5 +1,6 @@
 package com.almostreliable.unified.config;
 
+import com.almostreliable.unified.AlmostUnifiedPlatform;
 import com.almostreliable.unified.recipe.RecipeLink;
 import com.almostreliable.unified.utils.JsonCompare;
 import com.google.gson.JsonObject;
@@ -8,6 +9,7 @@ import net.minecraft.resources.ResourceLocation;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Set;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 public class DuplicationConfig extends Config {
@@ -15,11 +17,11 @@ public class DuplicationConfig extends Config {
 
     private final JsonCompare.CompareSettings defaultRules;
     private final LinkedHashMap<ResourceLocation, JsonCompare.CompareSettings> overrideRules;
-    private final Set<ResourceLocation> ignoreRecipeTypes;
-    private final Set<ResourceLocation> ignoreRecipes;
+    private final Set<Pattern> ignoreRecipeTypes;
+    private final Set<Pattern> ignoreRecipes;
     private final boolean strictMode;
 
-    public DuplicationConfig(JsonCompare.CompareSettings defaultRules, LinkedHashMap<ResourceLocation, JsonCompare.CompareSettings> overrideRules, Set<ResourceLocation> ignoreRecipeTypes, Set<ResourceLocation> ignoreRecipes, boolean strictMode) {
+    public DuplicationConfig(JsonCompare.CompareSettings defaultRules, LinkedHashMap<ResourceLocation, JsonCompare.CompareSettings> overrideRules, Set<Pattern> ignoreRecipeTypes, Set<Pattern> ignoreRecipes, boolean strictMode) {
         this.defaultRules = defaultRules;
         this.overrideRules = overrideRules;
         this.ignoreRecipeTypes = ignoreRecipeTypes;
@@ -28,7 +30,8 @@ public class DuplicationConfig extends Config {
     }
 
     public boolean shouldIgnoreRecipe(RecipeLink recipe) {
-        return ignoreRecipeTypes.contains(recipe.getType()) || ignoreRecipes.contains(recipe.getId());
+        return ignoreRecipeTypes.stream().anyMatch(pattern -> pattern.matcher(recipe.getType().toString()).matches()) ||
+               ignoreRecipes.stream().anyMatch(pattern -> pattern.matcher(recipe.getId().toString()).matches());
     }
 
     public JsonCompare.CompareSettings getCompareSettings(ResourceLocation type) {
@@ -48,14 +51,15 @@ public class DuplicationConfig extends Config {
 
         @Override
         public DuplicationConfig deserialize(JsonObject json) {
-            Set<ResourceLocation> ignoreRecipeTypes = deserializeResourceLocations(json,
+            var platform = AlmostUnifiedPlatform.INSTANCE.getPlatform();
+            Set<Pattern> ignoreRecipeTypes = deserializePatterns(json,
                     IGNORED_RECIPE_TYPES,
-                    Defaults.IGNORED_RECIPE_TYPES);
-            Set<ResourceLocation> ignoreRecipes = deserializeResourceLocations(json, IGNORED_RECIPES, List.of());
+                    Defaults.getIgnoredRecipeTypes(platform));
+            Set<Pattern> ignoreRecipes = deserializePatterns(json, IGNORED_RECIPES, List.of());
 
             JsonCompare.CompareSettings defaultRules = safeGet(() -> createCompareSet(json.getAsJsonObject(
                             DEFAULT_DUPLICATE_RULES)),
-                    defaultSettings());
+                    Defaults.getDefaultDuplicateRules(platform));
             LinkedHashMap<ResourceLocation, JsonCompare.CompareSettings> overrideRules = safeGet(() -> json
                     .getAsJsonObject(OVERRIDE_DUPLICATE_RULES)
                     .entrySet()
@@ -63,31 +67,10 @@ public class DuplicationConfig extends Config {
                     .collect(Collectors.toMap(entry -> new ResourceLocation(entry.getKey()),
                             entry -> createCompareSet(entry.getValue().getAsJsonObject()),
                             (a, b) -> b,
-                            LinkedHashMap::new)), defaultOverrides());
+                            LinkedHashMap::new)), Defaults.getDefaultDuplicateOverrides(platform));
             boolean strictMode = safeGet(() -> json.get(STRICT_MODE).getAsBoolean(), false);
 
             return new DuplicationConfig(defaultRules, overrideRules, ignoreRecipeTypes, ignoreRecipes, strictMode);
-        }
-
-        private JsonCompare.CompareSettings defaultSettings() {
-            JsonCompare.CompareSettings result = new JsonCompare.CompareSettings();
-            result.ignoreField("conditions");
-            result.ignoreField("group");
-            result.addRule("cookingtime", new JsonCompare.HigherRule());
-            result.addRule("energy", new JsonCompare.HigherRule());
-            result.addRule("experience", new JsonCompare.HigherRule());
-            return result;
-        }
-
-        private LinkedHashMap<ResourceLocation, JsonCompare.CompareSettings> defaultOverrides() {
-            JsonCompare.CompareSettings result = new JsonCompare.CompareSettings();
-            result.ignoreField("conditions");
-            result.ignoreField("group");
-            result.ignoreField("pattern");
-            result.ignoreField("key");
-            LinkedHashMap<ResourceLocation, JsonCompare.CompareSettings> resultMap = new LinkedHashMap<>();
-            resultMap.put(new ResourceLocation("minecraft", "crafting_shaped"), result);
-            return resultMap;
         }
 
         private JsonCompare.CompareSettings createCompareSet(JsonObject rules) {
@@ -100,8 +83,8 @@ public class DuplicationConfig extends Config {
         public JsonObject serialize(DuplicationConfig config) {
             JsonObject json = new JsonObject();
 
-            serializeResourceLocations(json, IGNORED_RECIPE_TYPES, config.ignoreRecipeTypes);
-            serializeResourceLocations(json, IGNORED_RECIPES, config.ignoreRecipes);
+            serializePatterns(json, IGNORED_RECIPE_TYPES, config.ignoreRecipeTypes);
+            serializePatterns(json, IGNORED_RECIPES, config.ignoreRecipes);
             json.add(DEFAULT_DUPLICATE_RULES, config.defaultRules.serialize());
             JsonObject overrides = new JsonObject();
             config.overrideRules.forEach((rl, compareSettings) -> {