+++ CHANGELOG.md
@@ -5,11 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [0.0.7] - 2022-09-16
+
+### Added
+- more materials and `storage_blocks/raw_{material}` tag to defaults
+  - back up your `unify.json` config and let it regenerate to get the new defaults or add them yourself
+- `fabric:conditions` to default duplicate ignore list
+
+### Changed
+- ignore lists in `unify.json` and `duplicates.json` now support regular expressions
+- default configs are now more platform specific
+
+### Fixed
+- a compat issue on Fabric when REI is present
+- items being hidden when they are the only entry in a tag
+
 ## [0.0.6] - 2022-09-02
 
 ### Added
 - more materials to defaults
-  - back up your `unify.json` config and let it regenerate to get the new default or add it yourself
+  - back up your `unify.json` config and let it regenerate to get the new defaults or add them yourself
 
 ### Fixed
 - stone strata detection for mods not following convention
@@ -19,7 +34,7 @@ and this project adheres to [Semantic Versioning].
 ### Added
 - `group` property to default ignored list
   - this ensures that crafting recipes with specific recipe book categories are correctly unified
-  - back up your `duplicates.json` config and let it regenerate to get the new default or add it yourself
+  - back up your `duplicates.json` config and let it regenerate to get the new defaults or add them yourself
 - `inputItems` and `outputItems` property to top level property scanning
   - allows unification for mods like `FTBIC`
 
@@ -70,6 +85,7 @@ Initial 1.19 release!
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
+[0.0.7]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.19-0.0.7-beta
 [0.0.6]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.19-0.0.6-beta
 [0.0.5]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.19-0.0.5-beta
 [0.0.4]: https://github.com/AlmostReliable/almostunified/releases/tag/v1.19-0.0.4-beta